requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
pytest
pytest-asyncio
python-multipart
aiofiles
jinja2